requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.1.0
mysql-connector-python>=8.0.33